KAFKA-13211: fix inMemory state store backward fetch issue
